\name{PropClust-package}
\alias{PropClust-package}
\alias{PropClust}
\docType{package}
\title{
Propensity Clustering and Decomposition.
}
\description{
This package implements two main functions PropensityDecomposition and PropensityClustering by finding the closest approximation to a matrix ADJ where ADJ_{ij}~A_{ab}a_i b_j with either L2 penalties or poisson penalties.
}
\details{
\tabular{ll}{
Package: \tab PropClust\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2013-02-20\cr
License: \tab GPL2 \cr
LazyLoad: \tab yes\cr
}
Index:
\preformatted{
propensityClustering    Propensity Clustering
CPBADecomposition       Cluster and Propensity-based Approximation decomposition for adajcency matrices
}
}

\author{
John Michael O. Ranola, Peter Langfelder, Steve Horvath, Kenneth Lange

Maintainer: Peter Langfelder <Peter.Langfelder@gmail.com>
}
\references{

Ranola JM, Langfelder P, Lange K, Horvath S (2013) Cluster and propensity based approximation of a network.
BMC Bioinformatics, in press.

Ranola et. al. (2010) A Poisson Model for Random Multigraphs. Bioinformatics 26(16):2004-2001.

}
\keyword{ cluster }
\keyword{ misc }
